Move tests to root with automatic PyCharm import refactoring. This will likely fail some tests
