+++ tests/envs/robotics/hand/test_manipulate_touch_sensors.py
@@ -3,7 +3,7 @@ import pickle
 import pytest
 
 from gym import envs
-from gym.tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
+from tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
 
 
 ENVIRONMENT_IDS = (
