+++ tests/vector/test_vector_env.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 
 from gym.spaces import Tuple
-from gym.tests.vector.utils import CustomSpace, make_env
+from tests.vector.utils import CustomSpace, make_env
 
 from gym.vector.async_vector_env import AsyncVectorEnv
 from gym.vector.sync_vector_env import SyncVectorEnv
