+++ tests/vector/utils.py
