+++ tests/wrappers/flatten_test.py
