+++ tests/wrappers/test_clip_action.py
