+++ tests/wrappers/test_flatten_observation.py
