+++ tests/wrappers/test_frame_stack.py
