+++ tests/wrappers/test_rescale_action.py
